refactor: Add AlertMessage component to login page
This commit adds the AlertMessage component to the login page in the `page.tsx` file. The AlertMessage component is imported from the newly created `src/components/general/alert-message.tsx` file. By including the AlertMessage component, we can display a success message to users after a successful registration. This change enhances the user experience and provides important feedback to the users.