refactor: Update eslint rules and package dependencies
This commit updates the eslint rules in the .eslintrc.json file to change the "@typescript-eslint/no-unused-vars" rule from "off" to "warn". This change will now show a warning for unused variables, helping to catch potential bugs and improve code quality.

Additionally, the commit updates the package.json file to upgrade the following dependencies:
- next-auth: from version 4.24.7 to 5.0.0-beta.19
- next-themes: from version ^0.3.0 to 0.3.0
- react-hook-form: from version ^7.52.0 to 7.52.0
- sonner: from version ^1.5.0 to 1.5.0

These updates ensure that the project is using the latest versions of these dependencies, which may include bug fixes, performance improvements, and new features.

Refactoring the eslint rules and updating the package dependencies improves code quality and keeps the project up-to-date with the latest best practices and improvements.