refactor: Delete user address functionality
This commit adds the `deleteUserAddress` function to the `address` directory, allowing users to delete their address from the database. The function uses the `prisma.userAddress.delete` method to delete the address with the specified ID. After successful deletion, the function triggers a revalidation of the `/shop/address` page. If the deletion fails, an error message is returned. This functionality improves the user experience by providing the ability to remove unwanted addresses.