refactor: Update address page to conditionally render address form or user address
This commit updates the address page component to conditionally render either the address form or the user address based on the presence of user addresses in the database. If there are no user addresses, the address form is displayed, allowing the user to add a new address. If there are existing user addresses, the user address component is displayed, showing the user's saved addresses and allowing them to select one for the checkout process. This change improves the user experience by providing a more streamlined and personalized address selection process.