refactor: Update discount calculation in cart components
This commit updates the calculation of discounts in the `CardCartItems` and `CardCartSummary` components. Instead of using a hardcoded discount value of 0.25, the components now check if the `item` object has a `discount` property and if it is greater than 0. If so, the discount is applied to the price calculation. This change ensures that discounts are correctly applied and displayed in the cart components.