feat: Add order schema and related types
This commit adds the `order-schema.ts` file to the `src/schema/order` directory. The `OrderSchema` object defines the structure of an order, including properties such as `id`, `subtotal`, `tax`, `total`, `itemsInOrder`, `isPaid`, `paidAt`, `transactionId`, `createdAt`, `updatedAt`, and `userId`. Additionally, the `ProductToOrderSchema` and `ProductsToOrderSchema` objects define the structure of the product-to-order relationship, including properties such as `productId` and `quantity`. These schema additions enable the implementation of order-related functionality in the application.