feat: Add discount converter utility function
Add a new utility function, `utilDiscountConverter`, to the `discount-converter.ts` file in the `src/utils/discount-converter` directory. This function calculates the discounted price and formats the currency for a given product's price and discount. It also determines if the product has a discount and calculates the percentage of the discount. This utility function enhances the functionality of the application by providing a convenient way to handle product discounts.