feat: Add base64 utility function and update utils index
Add a new utility function `toBase64` to convert a File object to base64 encoding. This function is added to the `base64.ts` file in the `utils/base64` directory. Additionally, update the `index.ts` file in the `utils` directory to export the `base64` utility function. These changes enhance the functionality of the application by providing a convenient way to convert files to base64 encoding.