feat: Add "publicId" field to ProductImage schema
Add a new field called "publicId" to the ProductImage schema in the `products-schema.ts` file. This field allows for storing the public ID of the image in Cloudinary for easier retrieval and management. This enhancement improves the functionality of the application by providing a way to associate Cloudinary public IDs with product images.