feat: Update ProductImage schema and types
Update the ProductImage schema and types to include the new "publicId" field. This field allows for storing the public ID of the image in Cloudinary for easier retrieval and management. This enhancement improves the functionality of the application by providing a way to associate Cloudinary public IDs with product images.