feat: Add deleteProductImage functionality and modal to UpdateProductImageForm
This commit adds the functionality to delete product images and a modal to the UpdateProductImageForm component. Users can now delete images associated with a product by clicking on a delete button displayed next to each image. This enhancement improves the user experience by providing a convenient way to remove unwanted images from products.