feat: Add discount column to Order and OrderItem models
This commit adds a new "discount" column to the "Order" and "OrderItem" models in the database schema. The "discount" column is of type DOUBLE PRECISION and has a default value of 0 for "Order" and 1 for "OrderItem". This enhancement allows for applying discounts to orders and order items, improving the functionality of the application.